refactor(cmd): restructure lifecycle commands into separate files
feat(cmd): add individual command files for start, stop, reload, status, and cleanup
refactor(cmd): move project-specific functions to project.go
test(cmd): implement comprehensive test suite for all commands
